encrypt-file/pyproject.toml
[project]
name = "encrypt-files"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "boto3[crt]>=1.42.44",
    "click>=8.3.1",
    "python-dotenv>=1.2.1",
]
